fix: PK 자료형 수정(Long->Int)

+++ src/main/kotlin/com/tik/server/repository/ResumeRepository.kt
@@ -3,6 +3,6 @@ package com.tik.server.repository
 import com.tik.server.entity.Resume
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ResumeRepository : JpaRepository<Resume, Long> {
+interface ResumeRepository : JpaRepository<Resume, Int> {
     fun findAllByMemberId(memberId: Int): List<Resume>
 }
